Renamed calculateViewWidths file for clarity, fixed speling, fixed navigation/lib exports to follow conventions
