+++ __tests__/commandMenu.test.tsx
@@ -3,7 +3,7 @@ import { NavigationProvider } from '@/navigation'
 import { render, screen } from '../setupTests'
 import { CommandMenu } from '@/components/CommandMenu'
 import userEvent from '@testing-library/user-event'
-import { initializeNavigationState } from '@/navigation/lib/initializeNavigationState'
+import { initializeNavigationState } from '@/navigation/lib'
 import { type NavigationActionType } from '@/types'
 import { type Dispatch } from 'react'
 import { useNavigation } from '@/hooks'
