+++ __tests__/viewFocus.test.tsx
@@ -8,7 +8,7 @@ import { useNavigation } from '@/navigation/hooks'
 import { render, screen } from '../setupTests'
 import { ViewFocus } from '@/components/ViewHeader/ViewFocus'
 import { type NavigationActionType } from '@/types'
-import { initializeNavigationState } from '@/navigation/lib/initializeNavigationState'
+import { initializeNavigationState } from '@/navigation/lib'
 
 
 jest.mock('@/navigation/hooks/useNavigation', () => ({
