+++ src/navigation/lib/index.tsx
@@ -1,3 +1,5 @@
 export { navigationReducer } from './navigationReducer'
 export { minimumSpaceRequired } from './minimumSpaceRequired'
 export { currentView } from './currentView'
+export { calculateViewWidths } from './calculateViewWidths'
+export { initializeNavigationState } from './initializeNavigationState'
